+++ src/reuse/report.py
@@ -14,14 +14,11 @@ from pathlib import Path
 from typing import Iterable, List, NamedTuple, Set
 from uuid import uuid4
 
-from license_expression import Licensing
-
 from . import __version__
-from ._util import _checksum
+from ._util import _LICENSING, _checksum
 from .project import Project
 
 _LOGGER = logging.getLogger(__name__)
-_LICENSING = Licensing()
 
 FileReportInfo = NamedTuple(
     "FileReportInfo",
@@ -187,6 +184,11 @@ class ProjectReport:  # pylint: disable=too-many-instance-attributes
                 project_report.bad_licenses.setdefault(name, set()).add(path)
 
         return project_report
+
+    @property
+    def used_licenses(self) -> Set[str]:
+        """Set of license identifiers that are found in file reports."""
+        return set(self.licenses) - self.unused_licenses
 
     @property
     def unused_licenses(self) -> Set[str]: