+++ tests/test_util.py
@@ -14,11 +14,10 @@ from pathlib import Path
 
 import pytest
 from boolean.boolean import ParseError
-from license_expression import LicenseSymbol, Licensing
+from license_expression import LicenseSymbol
 
 from reuse import _util
-
-_LICENSING = Licensing()
+from reuse._util import _LICENSING
 
 # pylint: disable=invalid-name
 git = pytest.mark.skipif(not _util.GIT_EXE, reason="requires git")
@@ -105,6 +104,29 @@ def test_find_root_in_git_repo(git_repository):
     result = _util.find_root()
 
     assert Path(result).absolute().resolve() == git_repository
+
+
+def test_make_copyright_line_simple():
+    """Given a simple statement, make it a copyright line."""
+    assert _util.make_copyright_line("hello") == "SPDX" "-Copyright: hello"
+
+
+def test_make_copyright_line_existing_spdx_copyright():
+    """Given a copyright line, do nothing."""
+    value = "SPDX" "-Copyright: hello"
+    assert _util.make_copyright_line(value) == value
+
+
+def test_make_copyright_line_existing_other_copyright():
+    """Given a non-SPDX copyright line, do nothing."""
+    value = "© hello"
+    assert _util.make_copyright_line(value) == value
+
+
+def test_make_copyright_line_multine_error():
+    """Given a multiline arguement, expect an error."""
+    with pytest.raises(RuntimeError):
+        _util.make_copyright_line("hello\nworld")
 
 
 # pylint: disable=unused-argument